name correction
