+++ fakturogen-comarch-connector/src/main/java/pl/fakturogen/comarch/connector/converter/ComarchConverter.java
@@ -1,6 +1,6 @@
 package pl.fakturogen.comarch.connector.converter;
 
-import pl.fakturogen.comarch.connector.exeption.converter.ComarchConverterException;
+import pl.fakturogen.comarch.connector.exception.converter.ComarchConverterException;
 
 public interface ComarchConverter<O> {
     O convert(String bodyString) throws ComarchConverterException;
