+++ fakturogen-comarch-connector/src/main/java/pl/fakturogen/comarch/connector/exception/ComarchConnectorException.java
@@ -1,4 +1,4 @@
-package pl.fakturogen.comarch.connector.exeption;
+package pl.fakturogen.comarch.connector.exception;
 
 public class ComarchConnectorException extends Exception {
     public ComarchConnectorException(String message) {
