+++ fakturogen-comarch-connector/src/main/java/pl/fakturogen/comarch/connector/exception/ComarchConverterException.java
@@ -1,8 +1,7 @@
-package pl.fakturogen.comarch.connector.exeption.converter;
-
-import pl.fakturogen.comarch.connector.exeption.ComarchConnectorException;
+package pl.fakturogen.comarch.connector.exception;
 
 public class ComarchConverterException extends ComarchConnectorException {
+
     public ComarchConverterException(String message) {
         super(message);
     }
