+++ fakturogen-comarch-connector/src/main/java/pl/fakturogen/comarch/connector/exception/CustomerNotFoundException.java
@@ -1,4 +1,4 @@
-package pl.fakturogen.comarch.connector.exeption;
+package pl.fakturogen.comarch.connector.exception;
 
 public class CustomerNotFoundException extends Exception {
     public CustomerNotFoundException(String message){
