+++ fakturogen-comarch-connector/src/main/java/pl/fakturogen/comarch/connector/exception/InvoiceNotFoundAdvice.java
@@ -1,0 +1,4 @@
+package pl.fakturogen.comarch.connector.exception;
+
+public class InvoiceNotFoundAdvice {
+}
