+++ fakturogen-comarch-connector/src/main/java/pl/fakturogen/comarch/connector/exception/InvoiceNotFoundException.java
@@ -1,4 +1,4 @@
-package pl.fakturogen.comarch.connector.exeption;
+package pl.fakturogen.comarch.connector.exception;
 
 public class InvoiceNotFoundException extends RuntimeException {
     public InvoiceNotFoundException(long id) {
