--- fakturogen-comarch-connector/src/main/java/pl/fakturogen/comarch/connector/exeption/ComarchConverterException.java
@@ -1,12 +1,0 @@
-package pl.fakturogen.comarch.connector.exeption;
-
-public class ComarchConverterException extends ComarchConnectorException {
-
-    public ComarchConverterException(String message) {
-        super(message);
-    }
-
-    public ComarchConverterException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
