--- fakturogen-comarch-connector/src/main/java/pl/fakturogen/comarch/connector/exeption/InvoiceNotFoundAdvice.java
@@ -1,4 +1,0 @@
-package pl.fakturogen.comarch.connector.exeption;
-
-public class InvoiceNotFoundAdvice {
-}
