+++ fakturogen-comarch-connector/src/main/java/pl/fakturogen/comarch/connector/services/ComarchProductService.java
@@ -2,7 +2,7 @@ package pl.fakturogen.comarch.connector.services;
 
 import org.springframework.stereotype.Service;
 import pl.fakturogen.comarch.connector.dto.ComarchProductDTO;
-import pl.fakturogen.comarch.connector.exeption.ComarchConnectorException;
+import pl.fakturogen.comarch.connector.exception.ComarchConnectorException;
 
 import java.util.List;
 import java.util.Optional;
