+++ fakturogen-comarch-connector/src/main/java/pl/fakturogen/comarch/connector/services/impl/ComarchInvoiceServiceDefault.java
@@ -3,10 +3,10 @@ package pl.fakturogen.comarch.connector.services.impl;
 import org.springframework.stereotype.Service;
 import pl.fakturogen.comarch.connector.connector.ComarchApiInvoiceConnector;
 import pl.fakturogen.comarch.connector.dto.ComarchInvoiceDTO;
-import pl.fakturogen.comarch.connector.exeption.ComarchConnectorException;
-import pl.fakturogen.comarch.connector.exeption.ComarchConverterException;
-import pl.fakturogen.comarch.connector.exeption.InvoiceNotFoundException;
-import pl.fakturogen.comarch.connector.exeption.InvoicesNotFoundException;
+import pl.fakturogen.comarch.connector.exception.ComarchConnectorException;
+import pl.fakturogen.comarch.connector.exception.ComarchConverterException;
+import pl.fakturogen.comarch.connector.exception.InvoiceNotFoundException;
+import pl.fakturogen.comarch.connector.exception.InvoicesNotFoundException;
 import pl.fakturogen.comarch.connector.mapper.ComarchInvoiceMapper;
 import pl.fakturogen.comarch.connector.model.ComarchInvoice;
 import pl.fakturogen.comarch.connector.services.ComarchInvoiceService;
