+++ fakturogen-comarch-connector/src/test/java/pl/fakturogen/comarch/connector/converter/ComarchCustomerConverterTest.java
@@ -5,7 +5,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import pl.fakturogen.comarch.connector.exeption.converter.ComarchConverterException;
+import pl.fakturogen.comarch.connector.exception.converter.ComarchConverterException;
 import pl.fakturogen.comarch.connector.model.ComarchAddress;
 import pl.fakturogen.comarch.connector.model.ComarchCustomer;
