+++ fakturogen-web/src/main/java/pl/fakturogen/web/controller/rest/TokenController.java
@@ -4,7 +4,7 @@ import org.springframework.boot.ApplicationArguments;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import pl.fakturogen.comarch.connector.connector.ComarchApiTokenConnector;
-import pl.fakturogen.comarch.connector.exeption.ComarchConnectorException;
+import pl.fakturogen.comarch.connector.exception.ComarchConnectorException;
 import pl.fakturogen.comarch.connector.model.ComarchToken;
 
 @RestController
